Add queries to bkAccountRoutes.js

+++ current-sea-back/routes/bkAccountRoutes.js
@@ -9,9 +9,16 @@ module.exports = function router() {
       const {
         userId, accountName,
       } = req.body;
-      db.query('INSERT INTO account_table(at_account_name, at_account_id, at_user_id) VALUES WHERE at_user_id = ? AND at_account_name = ?', [userId, accountName], (err) => {
-        if (err) {
-          throw err;
+      db.query('SELECT at_account_name from account_table where at_user_id = ? and at_account_name = ?', [userId, accountName], (err, result) => {
+        if (result.length === 0) {
+          db.query('INSERT INTO account_table(at_account_name, at_account_id, at_user_id) VALUES WHERE at_user_id = ? AND at_account_name = ?', [userId, accountName], (err, result) => {
+            if (err) {
+              throw err;
+            }
+            debug(result);
+          });
+        } else {
+          res.status(401).json({ message: 'Account already exists.' });
         }
       });
     });
@@ -23,12 +30,14 @@ module.exports = function router() {
 
   bkAccountRouter.route('/delete_account')
     .post((req, res) => {
-      const { userId } = req.params;
-      const { accountName } = req.params;
-      const { accountId } = req.params;
-      db.query('DELETE FROM account_table WHERE at_user_id = ? AND at_account_name = ? AND at_account_id = ?', [userId, accountName, accountId], (err) => {
-        if (err) {
-          throw err;
+      const { userId, accountId, accountName } = req.body;
+      db.query('SELECT account_name from account_table where at_user_id = ? AND at_account_name = ? AND at_account_id = ?', [userId, accountName, accountId], (err, result) => {
+        if (result.length !== 0) {
+          db.query('DELETE FROM account_table WHERE at_user_id = ? AND at_account_name = ? AND at_account_id = ?', [userId, accountName, accountId], (err) => {
+            if (err) {
+              throw err;
+            }
+          });
         }
       });
     });
